pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "azampay-sdk-anga"
version = "0.1.0"
description = "A dynamic Python SDK for AzamPay (AzamPay Python SDK used for native python, Django, Flask, FastApi e.t.c)"
readme = "README.md"
authors = [{ name = "Anganile Adam (Anga)", email = "twaloadam@gmail.com" }]
license = { text = "MIT" }
dependencies = [
    "requests",
    "python-dotenv"
]
requires-python = ">=3.7"

[project.urls]
"Homepage" = "https://github.com/tbwahacker/azampay-sdk-anga"
"Repository" = "https://github.com/tbwahacker/azampay-sdk-anga"
